point_cloud_editor/requirements.txt
PySide6==6.10.2
PyOpenGL==3.1.10
numpy
plyfile
scipy
